refactor(libs)!: move temperature into own lib
