+++ smarthome/Cargo.toml
@@ -19,6 +19,7 @@ empty_structs_with_brackets = "warn"
 [workspace]
 members = [
     # -- Application Libraries
+    "crates/libs/measure_units",
     "crates/libs/smarthome",
     "crates/libs/tcp-plug-socket",
     "crates/libs/tcp-plug-socket-async",
