+++ smarthome/crates/libs/measure_units/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "measure_units"
+version = "0.1.0"
+edition = "2021"
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+derive_more = "0.99.17"
+
+[lints]
+workspace = true
