+++ smarthome/crates/libs/udp-thermometer/src/lib.rs
@@ -4,12 +4,11 @@ use std::sync::{
     Arc, Mutex,
 };
 
+use measure_units::temperature::{self, Temperature};
 use smarthome::devices::Gauge;
-use temperature::Temperature;
 
 pub mod impl_async;
 pub mod impl_sync;
-pub mod temperature;
 
 pub type Result<T> = core::result::Result<T, Error>;
 pub type Error = Box<dyn std::error::Error>;
